added sendingevents v2
descripition as in commit before

+++ shopping-list-service/src/main/java/shopping/list/manager/shoppinglistservice/shoppingList/service/impl/ShoppingListDefaultService.java
@@ -1,7 +1,10 @@
 package shopping.list.manager.shoppinglistservice.shoppingList.service.impl;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import shopping.list.manager.shoppinglistservice.shoppingList.entity.ShoppingList;
 import shopping.list.manager.shoppinglistservice.shoppingList.repository.api.ShoppingListRepository;
+import shopping.list.manager.shoppinglistservice.shoppingList.service.api.ShoppingListEventPublisher;
 import shopping.list.manager.shoppinglistservice.shoppingList.service.api.ShoppingListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,10 +17,15 @@ import java.util.UUID;
 public class ShoppingListDefaultService implements ShoppingListService {
 
     private final ShoppingListRepository shoppingListRepository;
+    private final ShoppingListEventPublisher shoppingListEventPublisher;
+    private static final Logger logger = LoggerFactory.getLogger(ShoppingListDefaultService.class);
+
 
     @Autowired
-    public ShoppingListDefaultService(ShoppingListRepository shoppingListRepository) {
+    public ShoppingListDefaultService(ShoppingListRepository shoppingListRepository,
+                                      ShoppingListEventPublisher shoppingListEventPublisher) {
         this.shoppingListRepository = shoppingListRepository;
+        this.shoppingListEventPublisher = shoppingListEventPublisher;
     }
 
     @Override
@@ -25,6 +33,10 @@ public class ShoppingListDefaultService implements ShoppingListService {
         return shoppingListRepository.findById(id);
     }
 
+    @Override
+    public Optional<List<ShoppingList>> findAllByUserId(UUID userId){
+        return Optional.of(shoppingListRepository.findAllByUserId(userId));
+    }
 
     @Override
     public List<ShoppingList> findAll(){
@@ -43,7 +55,17 @@ public class ShoppingListDefaultService implements ShoppingListService {
 
     @Override
     public void delete(UUID id){
+        if (!shoppingListRepository.existsById(id)) {
+            logger.warn("Shopping list with id {} does not exist", id);
+            return;
+        }
         shoppingListRepository.deleteById(id);
+        shoppingListEventPublisher.notifyShoppingListDeleted(id);
+    }
+
+    @Override
+    public void deleteAllByUserId(UUID userId){
+        shoppingListRepository.deleteAllByUserId(userId);
     }
 
 }
